refactor: extract shared WebAdapterBase from GKT and wedge adapters
Fixes CI on the wedge PR:
- Prettier: remove stray blank lines in three touched files
- SonarCloud duplication: wedge adapter copied ~100 lines from the GKT
  adapter; both now extend a shared WebAdapterBase (scan callback
  management, one-shot single-tag capture, URL-key config, localStorage
  session persistence, key-preserving restart)
- SonarCloud coverage: base class tested directly, plus wedge tests for
  persistLog, getDeviceInfo and listener idempotence
