+++ src/hooks/useRfidScanning.ts
@@ -1,7 +1,6 @@
 import { adapter } from '@platform';
 import { useEffect, useRef, useCallback } from 'react';
 
-
 import type { NfcScanEvent } from '../platform/adapter';
 import { api, mapApiErrorToGerman, ApiError, formatRoomName } from '../services/api';
 import type { RfidScanResult, CurrentSession } from '../services/api';
