+++ src/platform/wedge/index.ts
@@ -11,8 +11,8 @@
  * colons (e.g. "F0:BC:E8:44"), so existing tag assignments keep matching.
  */
 
-import type { SessionSettings } from '../../services/sessionStorage';
-import type { NfcScanEvent, PlatformAdapter } from '../adapter';
+import type { PlatformAdapter } from '../adapter';
+import { WebAdapterBase } from '../webAdapterBase';
 
 /**
  * Maximum gap between two keystrokes (ms) to still count as scanner input.
@@ -39,12 +39,8 @@ export function normalizeWedgeUid(raw: string): string | null {
   return pairs ? pairs.join(':') : null;
 }
 
-class WedgeAdapter implements PlatformAdapter {
+class WedgeAdapter extends WebAdapterBase implements PlatformAdapter {
   readonly platform = 'wedge' as const;
-  private scanCallback: ((event: NfcScanEvent) => void) | null = null;
-  private cachedApiKey: string | null = null;
-  private scanCounter = 0;
-  private singleScanTimer: ReturnType<typeof setTimeout> | null = null;
 
   private buffer = '';
   private lastKeyAt = 0;
@@ -82,114 +78,9 @@ class WedgeAdapter implements PlatformAdapter {
     this.listenerAttached = true;
   }
 
-  private cancelPendingSingleScan(): void {
-    if (this.singleScanTimer) {
-      clearTimeout(this.singleScanTimer);
-      this.singleScanTimer = null;
-    }
-  }
-
-  async startScanning(onScan: (event: NfcScanEvent) => void): Promise<void> {
-    // Cancel any pending scanSingleTag timeout so its stale closure
-    // can't overwrite this newer callback when it fires.
-    this.cancelPendingSingleScan();
-    this.scanCallback = onScan;
-  }
-
-  async stopScanning(): Promise<void> {
-    this.cancelPendingSingleScan();
-    this.scanCallback = null;
-  }
-
-  async getServiceStatus(): Promise<{ is_running: boolean }> {
-    return { is_running: this.scanCallback !== null };
-  }
-
-  async scanSingleTag(
-    timeoutMs: number
-  ): Promise<{ success: boolean; tag_id?: string; error?: string }> {
-    // Same one-shot pattern as GKT: capture the next scan, then restore.
-    const previousCallback = this.scanCallback;
-    return new Promise(resolve => {
-      const oneShotCallback = (event: NfcScanEvent) => {
-        this.singleScanTimer = null;
-        clearTimeout(timer);
-        if (this.scanCallback === oneShotCallback) {
-          this.scanCallback = previousCallback;
-        }
-        resolve({ success: true, tag_id: event.tagId });
-      };
-
-      const timer = setTimeout(() => {
-        this.singleScanTimer = null;
-        if (this.scanCallback === oneShotCallback) {
-          this.scanCallback = previousCallback;
-        }
-        resolve({ success: false, error: 'Scan timed out' });
-      }, timeoutMs);
-
-      this.singleScanTimer = timer;
-      this.scanCallback = oneShotCallback;
-    });
-  }
-
-  async recoverScanner(): Promise<void> {
-    // No scanner recovery — the reader is a plain USB keyboard device
-  }
-
-  async getScannerStatus(): Promise<{
-    is_available: boolean;
-    last_error?: string;
-  }> {
-    // A keyboard wedge is invisible to the page until it types; assume available
-    return { is_available: true };
-  }
-
-  async loadConfig(): Promise<void> {
-    // No-op: wedge reads config synchronously from env vars / URL params
-  }
-
-  getApiBaseUrl(): string {
-    return (import.meta.env.VITE_API_BASE_URL as string) ?? 'https://api.moto-app.de';
-  }
-
-  getDeviceApiKey(): string {
-    const params = new URLSearchParams(window.location.search);
-    const key = params.get('key');
-    if (!key) {
-      if (this.cachedApiKey) {
-        return this.cachedApiKey;
-      }
-      throw new Error('DEVICE_API_KEY not found in URL. Expected ?key=...');
-    }
-    this.cachedApiKey = key;
-    return key;
-  }
-
-  async saveSessionSettings(settings: SessionSettings): Promise<void> {
-    localStorage.setItem('pyreportal_session', JSON.stringify(settings));
-  }
-
-  async loadSessionSettings(): Promise<SessionSettings | null> {
-    const data = localStorage.getItem('pyreportal_session');
-    return data ? (JSON.parse(data) as SessionSettings) : null;
-  }
-
-  async clearLastSession(): Promise<void> {
-    localStorage.removeItem('pyreportal_session');
-  }
-
   async persistLog(entry: string): Promise<void> {
     // eslint-disable-next-line no-console
     console.log('[PyrePortal]', entry);
-  }
-
-  async restartApp(): Promise<void> {
-    if (this.cachedApiKey) {
-      window.location.href = `${window.location.origin}/?key=${encodeURIComponent(this.cachedApiKey)}`;
-    } else {
-      window.location.reload();
-    }
   }
 
   getDeviceInfo(): { platform: 'wedge'; version: string } {
